requirements.txt
steam==1.0.2
stem==1.8.0
requests==2.23.0
colorama==0.4.3
termcolor==1.1.0
beautifulsoup4==4.9.1
SocksiPy_branch==1.01
